change int to double

+++ helloworld.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-  int a;
-  int b;
+  double a;
+  double b;
   cin >> a >> b;
   cout << a + b << endl;
